Further divided project into subpackages.
During this process, I had to change the permissions of severa member
variables in the Ship class - the variables were all package-private, so
I made them public. This is not great, but since they were previously
package-private in the default package, it is still an improvement over
the original code.